# crates/full-node/sov-api-spec/openapi-v3.yaml
...
openapi: 3.0.2
info:
  title: Sovereign Full Node API
  version: 0.1.0
  description: |-
    This is the REST API for the Sovereign SDK full node.
  contact:
    name: Sovereign Labs
    url: https://github.com/Sovereign-Labs/sovereign-sdk
    email: info@sovereign.xyz
  license:
    name: Sovereign Labs Community Software License
    url: https://github.com/Sovereign-Labs/sovereign-sdk/blob/nightly/LICENSE.md
servers:
  - url: http://localhost:12346
    description: Localhost
security: []
paths:
...
  /rollup/schema:
    get:
      summary: Get schema
      description: >
        Returns the current schema used by the rollup.

        This information can be used by web3 SDK and external tools as a description to serialize objects
        as well as determine if the version of the rollup has changed.
      operationId: rollup_schema
      tags:
        - Rollup
      responses:
        "200":
          $ref: "#/components/responses/Schema"
        "400":
          $ref: "#/components/responses/ApiErrorResponse"
...
components:
...
  responses:
...
    Schema:
      description: Schema of the rollup.
      content:
        application/json:
          schema:
            type: object
            properties:
              schema:
                type: object
                additionalProperties: true
              chain_hash:
                $ref: "#/components/schemas/HexString"
            required:
              - schema
              - chain_hash
...
    ApiErrorResponse:
      description: The request sent was invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    HexString:
      type: string
      pattern: ^0x[a-fA-F0-9]+$
      description: Hex-encoded byte sequence
      example: "0x71c7656ec7ab88b098defb751b7401b5f6d8976f"
...
    ApiError:
      type: object
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code related to this error
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - status
        - message
        - details
...